refactor OSSupport
